+++ tests/unit/test_utils.py
@@ -15,8 +15,16 @@
 #
 
 from unittest import TestCase
+try:
+    from unittest import mock
+except ImportError:
+    import mock
 
-from sparkly.utils import parse_schema
+from pyspark import StorageLevel
+from pyspark.sql import DataFrame
+from pyspark.sql import types as T
+
+from sparkly.utils import lru_cache, parse_schema
 
 
 class TestParseSchema(TestCase):
@@ -44,3 +52,45 @@ class TestParseSchema(TestCase):
 
     def assert_parsed_properly(self, schema):
         self.assertEqual(parse_schema(schema).simpleString(), schema)
+
+
+class TestLruCache(TestCase):
+    def test_caching(self):
+        df = mock.MagicMock(spec=DataFrame)
+
+        called = [0]
+
+        @lru_cache(storage_level=StorageLevel.DISK_ONLY)
+        def func(*args, **kwargs):
+            called[0] += 1
+            return df
+
+        func()
+        df.persist.assert_called_once_with(StorageLevel.DISK_ONLY)
+        self.assertEqual(df.unpersist.mock_calls, [])
+        self.assertEqual(called[0], 1)
+
+        cached_df = func()
+        self.assertEqual(cached_df, df)
+        self.assertEqual(called[0], 1)
+
+    def test_eviction(self):
+        first_df = mock.MagicMock(spec=DataFrame)
+        second_df = mock.MagicMock(spec=DataFrame)
+
+        @lru_cache(maxsize=1, storage_level=StorageLevel.DISK_ONLY)
+        def func(uid):
+            if uid == 'first':
+                return first_df
+            else:
+                return second_df
+
+        func('first')
+        first_df.persist.assert_called_once_with(StorageLevel.DISK_ONLY)
+        self.assertEqual(first_df.unpersist.mock_calls, [])
+
+        func('second')
+        first_df.persist.assert_called_once_with(StorageLevel.DISK_ONLY)
+        first_df.unpersist.assert_called_once_with()
+        second_df.persist.assert_called_once_with(StorageLevel.DISK_ONLY)
+        self.assertEqual(second_df.unpersist.mock_calls, [])